ENH: Allowing the usage of float as input image

+++ Modules/Filtering/TextureFeatures/test/RunLengthTextureFeaturesImageFilterTest.cxx
@@ -46,7 +46,7 @@ RunLengthTextureFeaturesImageFilterTest(int argc, char * argv[])
   const unsigned int VectorComponentDimension = 10;
 
   // Declare types
-  typedef int                                                             InputPixelType;
+  typedef float                                                           InputPixelType;
   typedef float                                                           OutputPixelComponentType;
   typedef itk::Vector<OutputPixelComponentType, VectorComponentDimension> OutputPixelType;
